requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==5.1.0
celery==5.2.6
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.13
django-celery-beat==2.2.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
kombu==5.2.4
lxml==4.8.0
Markdown==3.3.7
oauth2client==4.1.3
oauthlib==3.2.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.8.0
